# docs/swagger.yaml
basePath: /lending-engine
definitions:
...
  account.ConfirmAccountAdminRequest:
    properties:
      accountId:
        example: 1
        type: integer
    type: object
...
  account.GetAccountAdminResponse:
    properties:
      accountId:
        example: 1
        type: integer
      accountNumber:
        example: "11111111"
        type: string
      document:
        items:
          $ref: '#/definitions/account.Document'
        type: array
      email:
        example: k.apiwattanawong@gmail.com
        type: string
      firstName:
        example: Khanapat
        type: string
      isVerify:
        example: true
        type: boolean
      lastName:
        example: Apiwattanawong
        type: string
      password:
        example: <bcrypt>
        type: string
      phone:
        example: "0811111111"
        type: string
      status:
        example: PENDING
        type: string
      termCondition:
        example: 1.0.0
        type: string
    type: object
...
  lending.ConfirmLoanAdminRequest:
    properties:
      id:
        example: 1
        type: integer
    type: object
...
  response.ErrResponse:
    properties:
      code:
        example: 400
        type: integer
      description:
        example: Please contact administrator for more information.
        type: string
      error:
        example: '''<Field>'' must be REQUIRED field but the input is '''''
        type: string
      title:
        example: Cannot register public key.
        type: string
    type: object
  response.Response:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      title:
        example: Register key success.
        type: string
    type: object
host: localhost:9090
info:
  contact:
    email: k.apiwattanawong@gmail.com
    name: K.Apiwattanawong
    url: http://www.swagger.io/support
  description: Lending finance for ICFIN company.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lending Financial Services
  version: "1.0"
paths:
  /admin/account:
    get:
      consumes:
      - application/json
      description: get account by account id or email
      parameters:
      - description: Account ID
        in: query
        name: accountId
        type: integer
      - description: Email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/account.GetAccountAdminResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get Account Admin
      tags:
      - Admin
  /admin/account/confirm:
    post:
      consumes:
      - application/json
      description: confirm account by account id
      parameters:
      - description: request body to confirm account
        in: body
        name: ConfirmAccountAdmin
        required: true
        schema:
          $ref: '#/definitions/account.ConfirmAccountAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Confirm Account Admin
      tags:
      - Admin
  /admin/account/document:
    put:
      consumes:
      - application/json
      description: update account document by account id and document id
      parameters:
      - description: request body to update account document
        in: body
        name: UpdateAccountDocumentAdmin
        required: true
        schema:
          $ref: '#/definitions/account.UpdateAccountDocumentAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Update Account document Admin
      tags:
      - Admin
  /admin/account/reject:
    post:
      consumes:
      - application/json
      description: reject account by account id
      parameters:
      - description: request body to reject account
        in: body
        name: RejectAccountAdmin
        required: true
        schema:
          $ref: '#/definitions/account.RejectAccountAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Reject Account Admin
      tags:
      - Admin
  /admin/contract:
    get:
      consumes:
      - application/json
      description: get loan by contract id or account id
      parameters:
      - description: Contract ID
        in: query
        name: contractId
        type: string
      - description: Account ID
        in: query
        name: accountId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/lending.Contract'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get Loan Admin
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: confirm loan contract by account id
      parameters:
      - description: request body to confirm loan contract
        in: body
        name: ConfirmLoanAdmin
        required: true
        schema:
          $ref: '#/definitions/lending.ConfirmLoanAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Confirm Loan Admin
      tags:
      - Admin
  /admin/deposit/confirm:
    post:
      consumes:
      - application/json
      description: confirm deposit transaction by account id
      parameters:
      - description: request body to confirm deposit
        in: body
        name: ConfirmDepositAdmin
        required: true
        schema:
          $ref: '#/definitions/lending.ConfirmDepositAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Confirm Deposit Admin
      tags:
      - Admin
  /admin/deposit/reject:
    post:
      consumes:
      - application/json
      description: reject deposit transaction by account id
      parameters:
      - description: request body to reject deposit
        in: body
        name: RejectDepositAdmin
        required: true
        schema:
          $ref: '#/definitions/lending.RejectDepositAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Reject Deposit Admin
      tags:
      - Admin
  /admin/documentInfo:
    get:
      consumes:
      - application/json
      description: get all of document info
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/account.DocumentInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get Document Info Admin
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: create new document info
      parameters:
      - description: request body to create document info
        in: body
        name: CreateDocumentInfo
        required: true
        schema:
          $ref: '#/definitions/account.CreateDocumentInfoAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/account.CreateDocumentInfoAdminResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Create Document Info Admin
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: update document type by document id
      parameters:
      - description: request body to update document info
        in: body
        name: UpdateDocumentInfo
        required: true
        schema:
          $ref: '#/definitions/account.UpdateDocumentInfoAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Update Document Info Admin
      tags:
      - Admin
  /admin/interest:
    get:
      consumes:
      - application/json
      description: get all of interest term
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/lending.InterestTerm'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get Interest Term Admin
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: create new interest rate
      parameters:
      - description: request body to create interest term
        in: body
        name: CreateInterestTerm
        required: true
        schema:
          $ref: '#/definitions/lending.CreateInterestTermAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lending.CreateInterestTermAdminResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Create Interest Term Admin
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: update interest rate by interest code
      parameters:
      - description: request body to update interest term
        in: body
        name: UpdateInterestTerm
        required: true
        schema:
          $ref: '#/definitions/lending.UpdateInterestTermAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Update Interest Term Admin
      tags:
      - Admin
  /admin/repay:
    get:
      consumes:
      - application/json
      description: get repayment by id, contract id and account id
      parameters:
      - description: ID
        in: query
        name: id
        type: integer
      - description: Contract ID
        in: query
        name: contractId
        type: integer
      - description: Account ID
        in: query
        name: accountId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/lending.RepayTransaction'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get Repay Admin
      tags:
      - Admin
  /admin/repay/confirm:
    post:
      consumes:
      - application/json
      description: confirm repayment by account id
      parameters:
      - description: request body to confirm repay
        in: body
        name: ConfirmRepayAdmin
        required: true
        schema:
          $ref: '#/definitions/lending.ConfirmRepayAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Confirm Repay Admin
      tags:
      - Admin
  /admin/repay/reject:
    post:
      consumes:
      - application/json
      description: reject repayment by account id
      parameters:
      - description: request body to reject repay
        in: body
        name: RejectRepayAdmin
        required: true
        schema:
          $ref: '#/definitions/lending.RejectRepayAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Reject Repay Admin
      tags:
      - Admin
  /admin/wallet-transaction:
    get:
      consumes:
      - application/json
      description: get wallet transaction by id, account id, address or txn type
      parameters:
      - description: Transaction ID
        in: query
        name: Id
        type: integer
      - description: Account ID
        in: query
        name: accountId
        type: integer
      - description: Address
        in: query
        name: address
        type: string
      - description: Transaction Type
        in: query
        name: txnType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lending.WalletTransaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get Wallet Transaction Admin
      tags:
      - Admin
  /admin/withdraw/confirm:
    post:
      consumes:
      - application/json
      description: confirm withdraw transaction by account id
      parameters:
      - description: request body to confirm withdraw
        in: body
        name: ConfirmWithdrawAdmin
        required: true
        schema:
          $ref: '#/definitions/lending.ConfirmWithdrawAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Confirm Withdraw Admin
      tags:
      - Admin
  /admin/withdraw/reject:
    post:
      consumes:
      - application/json
      description: reject withdraw transaction by account id
      parameters:
      - description: request body to reject withdraw
        in: body
        name: RejectWithdrawAdmin
        required: true
        schema:
          $ref: '#/definitions/lending.RejectWithdrawAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Reject Withdraw Admin
      tags:
      - Admin
  /borrow:
    post:
      consumes:
      - application/json
      description: borrow loan
      parameters:
      - description: reference number.
        in: header
        name: ReferenceNo
        required: true
        type: string
      - description: one time password.
        in: header
        name: OTP
        required: true
        type: string
      - description: request body to borrow loan
        in: body
        name: BorrowLoan
        required: true
        schema:
          $ref: '#/definitions/lending.BorrowLoanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lending.BorrowLoanResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Borrow Loan
      tags:
      - Lending
  /contract:
    get:
      consumes:
      - application/json
      description: get user's loan contract by accountId
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/lending.Contract'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Contract Loan
      tags:
      - Lending
  /credit:
    get:
      consumes:
      - application/json
      description: get user's credit available by accountId
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lending.GetCreditAvailableResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Credit Available
      tags:
      - Lending
  /deposit:
    post:
      consumes:
      - application/json
      description: submit deposit transaction
      parameters:
      - description: request body to submit deposit
        in: body
        name: SubmitDeposit
        required: true
        schema:
          $ref: '#/definitions/lending.SubmitDepositRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lending.SubmitDepositResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Submit Deposit
      tags:
      - Lending
  /login:
    post:
      consumes:
      - application/json
      description: login by user & password
      parameters:
      - description: request body to login account
        in: body
        name: Login
        required: true
        schema:
          $ref: '#/definitions/account.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/account.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Login
      tags:
      - Account
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"